fix: Update MessageAudio component to use TTSEndpoints enum
