+++ client/src/components/Chat/Messages/MessageAudio.tsx
@@ -3,14 +3,15 @@ import { memo } from 'react';
 import type { TMessageAudio } from '~/common';
 import { BrowserTTS, EdgeTTS, ExternalTTS } from '~/components/Audio/TTS';
 import { useGetAudioSettings } from '~/hooks';
+import { TTSEndpoints } from '~/common';
 
 function MessageAudio(props: TMessageAudio) {
   const { textToSpeechEndpoint } = useGetAudioSettings();
 
   const TTSComponents = {
-    edge: EdgeTTS,
-    external: ExternalTTS,
-    default: BrowserTTS,
+    [TTSEndpoints.edge]: EdgeTTS,
+    [TTSEndpoints.browser]: BrowserTTS,
+    [TTSEndpoints.external]: ExternalTTS,
   };
 
   const SelectedTTS = TTSComponents[textToSpeechEndpoint];
